pyproject.toml
[project]
name = "db2-til-bq"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "db-dtypes>=1.4.3",
    "google-auth>=2.40.3",
    "google-cloud-bigquery>=3.34.0",
    "google-cloud-bigquery-storage>=2.32.0",
    "ibm-db>=3.2.6",
    "ibm-db-sa>=0.4.1",
    "pandas>=2.3.0",
    "pandas-gbq>=0.29.2",
    "pyyaml>= 6.0.3"
]

[tool.ruff]
exclude = [".venv", "explore/"]

[dependency-groups]
dev = [
    "python-dotenv>=1.1.1",
    "ruff>=0.14.9",
    "toml-to-requirements>=0.3.0",
]
